Fix timing issue in Ninox acquisition
Was able to fix the time lag in the image acquisition for the Ninox in MM. Running as a sequence seemed to populate the image buffer but was unable to synchronize the reading of the images. This caused a delay where frames would subsequently appear later and later in the image sequence than they were supposed to. Ended up switching to a for loop and taking single snaps. It wasn't working at first (same image for every frame) but a reset fixed it. However, switching the state of the Arduino to trigger ThorCam is still slow, which limits the maximum frame rate.